Add HomeController Rename Package spring to workman Update DB Connect Info localhost to workman
