--- src/main/java/com/kh/spring/collabo/model/dao/CollaboDao.java
@@ -1,5 +1,0 @@
-package com.kh.spring.collabo.model.dao;
-
-public interface CollaboDao {
-
-}
