--- src/main/java/com/kh/spring/collabo/model/service/CollaboService.java
@@ -1,5 +1,0 @@
-package com.kh.spring.collabo.model.service;
-
-public interface CollaboService {
-
-}
