--- src/main/java/com/kh/spring/dm/controller/EmpController.java
@@ -1,5 +1,0 @@
-package com.kh.spring.dm.controller;
-
-public class EmpController {
-
-}
