--- src/main/java/com/kh/spring/dm/model/vo/Employee.java
@@ -1,20 +1,0 @@
-package com.kh.spring.dm.model.vo;
-
-import java.sql.Date;
-
-public class Employee {
-  private String empId;
-  private String empNo;
-  private String email;
-  private String phone;
-  private String deptCode;
-  private String jobCode;
-  private String salLevel;
-  private int salary;
-  private int bonus;
-  private String managerId;
-  private Date hireDate;
-  private Date entDate;
-  private String entYn;
-
-}
