+++ src/main/java/com/kh/workman/collabo/controller/CollaboController.java
@@ -1,11 +1,11 @@
-package com.kh.spring.collabo.controller;
+package com.kh.workman.collabo.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
-import com.kh.spring.collabo.model.service.CollaboService;
+import com.kh.workman.collabo.model.service.CollaboService;
 
 @Controller
 public class CollaboController {
