+++ src/main/java/com/kh/workman/collabo/model/dao/CollaboDao.java
@@ -1,0 +1,5 @@
+package com.kh.workman.collabo.model.dao;
+
+public interface CollaboDao {
+
+}
