+++ src/main/java/com/kh/workman/collabo/model/service/CollaboService.java
@@ -1,0 +1,5 @@
+package com.kh.workman.collabo.model.service;
+
+public interface CollaboService {
+
+}
