+++ src/main/java/com/kh/workman/collabo/model/service/CollaboServiceImpl.java
@@ -1,10 +1,10 @@
-package com.kh.spring.collabo.model.service;
+package com.kh.workman.collabo.model.service;
 
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.kh.spring.collabo.model.dao.CollaboDao;
+import com.kh.workman.collabo.model.dao.CollaboDao;
 
 @Service
 public class CollaboServiceImpl implements CollaboService {
