+++ src/main/java/com/kh/workman/collabo/model/vo/CollaboCard.java
@@ -1,4 +1,4 @@
-package com.kh.spring.collabo.model.vo;
+package com.kh.workman.collabo.model.vo;
 
 import java.sql.Date;
 
@@ -7,10 +7,10 @@ import lombok.ToString;
 
 @Data
 @ToString
-public class CollaboList {
+public class CollaboCard {
+	private int cardNo;
 	private int listNo;
-	private int collaboNo;
-	private String title;
 	private int writer;
+	private String content;
 	private Date regdate;
 }
