+++ src/main/java/com/kh/workman/common/ArrayTypeHandler.java
@@ -1,4 +1,4 @@
-package com.kh.spring.common;
+package com.kh.workman.common;
 
 import java.sql.CallableStatement;
 import java.sql.PreparedStatement;
