+++ src/main/java/com/kh/workman/common/interceptor/LoggerInterceptor.java
@@ -1,4 +1,4 @@
-package com.kh.spring.common.interceptor;
+package com.kh.workman.common.interceptor;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
