+++ src/main/java/com/kh/workman/home/controller/HomeController.java
@@ -1,12 +1,13 @@
-package com.kh.spring.dm.controller;
+package com.kh.workman.home.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 public class HomeController {
-  @RequestMapping("homeView")
-  public String viewHome() {
-    return "mainView";
-  }
+	
+	@RequestMapping("/")
+	public String home() {
+		return "index";
+	}
 }
