+++ src/main/resources/mybatis-config.xml
@@ -4,7 +4,6 @@
 <!-- connection info는 root-context에서 설정됨. 여기서는 type alias 등 설정 -->
 <configuration>
   <typeAliases>
-    <typeAlias type="com.kh.spring.dm.model.vo.Employee" alias="emp" />
-    <typeAlias type="com.kh.spring.common.ArrayTypeHandler" alias="arrType" />
+    <typeAlias type="com.kh.workman.common.ArrayTypeHandler" alias="arrType" />
  </typeAliases>
 </configuration>
