+++ src/main/webapp/WEB-INF/spring/appServlet/servlet-context.xml
@@ -21,13 +21,13 @@
     <beans:property name="suffix" value=".jsp" />
   </beans:bean>
   
-  <context:component-scan base-package="com.kh.spring" />
+  <context:component-scan base-package="com.kh.workman" />
 
   <!-- 인터셉터 등록하기 -->
   <interceptors>
     <interceptor>
       <mapping path="/**" />
-      <beans:bean id="loggerInterceptor" class="com.kh.spring.common.interceptor.LoggerInterceptor" />
+      <beans:bean id="loggerInterceptor" class="com.kh.workman.common.interceptor.LoggerInterceptor" />
     </interceptor>
   </interceptors>
