+++ src/main/webapp/WEB-INF/spring/root-context.xml
@@ -23,9 +23,9 @@
   -->
   <bean id="dataSource" class="org.apache.commons.dbcp.BasicDataSource" destroy-method="close">
     <property name="driverClassName" value="oracle.jdbc.driver.OracleDriver"/>
-    <property name="url" value="jdbc:oracle:thin:@localhost:1521:xe"/>
-    <property name="username" value="kh"/>
-    <property name="password" value="kh"/>
+    <property name="url" value="jdbc:oracle:thin:@rclass.iptime.org:1521:xe"/>
+    <property name="username" value="workman"/>
+    <property name="password" value="workman"/>
   </bean>
 
   <!-- Mybatis에서 SqlSession 객체를 생성하는 FactoryBean 을 springbean으로 등록 
@@ -41,7 +41,7 @@
     <constructor-arg index="0" ref="sqlSessionFactoryBean"/>
   </bean> 
   
-  <bean id="collaboHandler" class="com.kh.spring.collabo.controller.CollaboHandler"/>
+  <bean id="collaboHandler" class="com.kh.workman.collabo.controller.CollaboHandler"/>
   <websocket:handlers>
   	<websocket:mapping handler="collaboHandler" path="/collabo/soc"/>
   	<websocket:sockjs/>
