+++ src/test/resources/log4j.xml
@@ -11,8 +11,8 @@
 	</appender>
 	
 	<!-- Application Loggers -->
-	<logger name="com.kh.spring">
-		<level value="info" />
+	<logger name="com.kh.workman">
+		<level value="debug" />
 	</logger>
 	
 	<!-- 3rdparty Loggers -->
